+++ .eslintignore
@@ -8,11 +8,6 @@ scripts/
 stacks-blockchain/
 src/schemas/*
 docs/
-#client/
-#src/tests/
-src/tests-rosetta/
-src/tests-rosetta-cli/
-src/tests-bns/
 client/src/
 config/
 utils/src/
