+++ .github/workflows/ci.yml
@@ -102,9 +102,9 @@ jobs:
       - name: Setup integration environment
         run: |
           sudo ufw disable
-          mkdir -p src/tests-event-replay/.tmp/local/
-          sudo chown 999:999 src/tests-event-replay/.tmp/local/
-          sudo chmod -R 777 src/tests-event-replay/.tmp/local/
+          mkdir -p tests/event-replay/.tmp/local/
+          sudo chown 999:999 tests/event-replay/.tmp/local/
+          sudo chmod -R 777 tests/event-replay/.tmp/local/
           npm run devenv:deploy -- -d
           npm run devenv:logs -- --no-color &> docker-compose-logs.txt &
